requirements.txt
django==5.0.2
kinde-python-sdk==1.2.6
python-dotenv==1.0.0